+++ meta/and.rb
@@ -7,8 +7,6 @@ Mutant::Meta::Example.add :and do
   mutation 'true'
   mutation 'false'
   mutation 'true or false'
-  mutation 'true and nil'
-  mutation 'nil and false'
   mutation 'false and false'
   mutation 'true and true'
   mutation '!true and false'