+++ meta/case.rb
@@ -23,12 +23,6 @@ Mutant::Meta::Example.add :case do
     else
     end
   RUBY
-  mutation <<-RUBY
-    case
-    when nil
-    else
-    end
-  RUBY
 end
 
 # rubocop:disable Metrics/BlockLength