+++ meta/if.rb
@@ -28,11 +28,9 @@ Mutant::Meta::Example.add :if do
 
   # mutation of if body
   mutation 'if condition; false; else false; end'
-  mutation 'if condition; nil;   else false; end'
 
   # mutation of else body
   mutation 'if condition; true;  else true;  end'
-  mutation 'if condition; true;  else nil;   end'
 end
 
 Mutant::Meta::Example.add :if do
@@ -41,7 +39,6 @@ Mutant::Meta::Example.add :if do
   singleton_mutations
   mutation 'if !condition; true;  end'
   mutation 'if condition;  false; end'
-  mutation 'if condition;  nil;   end'
   mutation 'if true;       true;  end'
   mutation 'if false;      true;  end'
   mutation 'if nil;        true;  end'
@@ -57,7 +54,6 @@ Mutant::Meta::Example.add :if do
   mutation 'unless true;       true;  end'
   mutation 'unless false;      true;  end'
   mutation 'unless condition;  false; end'
-  mutation 'unless condition;  nil;   end'
   mutation 'if     condition;  true;  end'
   mutation 'true'
 end
@@ -68,7 +64,6 @@ Mutant::Meta::Example.add :if do
   singleton_mutations
   mutation 'false if /foo/'
   mutation 'true if //'
-  mutation 'nil if /foo/'
   mutation 'true if true'
   mutation 'true if false'
   mutation 'true if nil'