+++ meta/regexp.rb
@@ -58,7 +58,6 @@ Mutant::Meta::Example.add :regexp do
 
   singleton_mutations
   mutation 'false if /foo/'
-  mutation 'nil if /foo/'
   mutation 'true if true'
   mutation 'true if false'
   mutation 'true if nil'